+++ UnitTests/BraintreePayPalTests/BTPayPalCheckoutRequest_Tests.swift
@@ -87,6 +87,7 @@ class BTPayPalCheckoutRequest_Tests: XCTestCase {
         request.billingAgreementDescription = "description"
         request.userAction = .payNow
         request.userAuthenticationEmail = "fake@email.com"
+        request.userPhoneNumber = "4087463271"
 
         let shippingAddress = BTPostalAddress()
         shippingAddress.streetAddress = "123 Main"
@@ -114,6 +115,7 @@ class BTPayPalCheckoutRequest_Tests: XCTestCase {
         XCTAssertEqual(parameters["recipient_name"] as? String, "Recipient")
         XCTAssertEqual(parameters["payer_email"] as? String, "fake@email.com")
         XCTAssertEqual(parameters["request_billing_agreement"] as? Bool, true)
+        XCTAssertEqual(parameters["phone"] as? String, "4087463271")
 
         guard let billingAgreementDetails = parameters["billing_agreement_details"] as? [String : String] else {
             XCTFail()